+++ src/expressions/boolean.js
@@ -8,7 +8,7 @@ class LimeExpressionBoolean extends LimeExpression {
   // Constructor
   constructor(lime, value) {
     // Super from expression class
-    super(lime, { name: 'boolean', shorthand: 'bool', isSimple: true });
+    super(lime, { name: 'boolean', isSimple: true });
 
     // Build expression
     this.raw = value;